Update base test case
Travis already clears the db, so we can skip that in the setUp stage.

+++ src/tests/ggrc/__init__.py
@@ -22,12 +22,13 @@ logging.disable(logging.CRITICAL)
 
 class TestCase(BaseTestCase):
   def setUp(self):
-    db.engine.execute("SET FOREIGN_KEY_CHECKS=0")
-    for table in reversed(db.metadata.sorted_tables):
-      if table.name not in ("test_model", "roles", "notification_types", "object_types"):
-        db.engine.execute(table.delete())
-    db.engine.execute("SET FOREIGN_KEY_CHECKS=1")
-    db.session.commit()
+    if not os.environ.get('TRAVIS', False):
+      db.engine.execute("SET FOREIGN_KEY_CHECKS=0")
+      for table in reversed(db.metadata.sorted_tables):
+        if table.name not in ("test_model", "roles", "notification_types", "object_types"):
+          db.engine.execute(table.delete())
+      db.engine.execute("SET FOREIGN_KEY_CHECKS=1")
+      db.session.commit()
 
     # if getattr(settings, 'MEMCACHE_MECHANISM', False) is True:
     #   from google.appengine.api import memcache
